% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_quick_map.R
\name{acs_quick_map}
\alias{acs_quick_map}
\title{Quickly create a choropleth sketch}
\usage{
acs_quick_map(
  .data,
  name = name,
  value = value,
  level = c("town", "neighborhood", "tract"),
  city = NULL,
  n = 5,
  palette = "GnBu",
  title = NULL,
  ...
)
}
\arguments{
\item{.data}{A data frame containing data by geography.}

\item{name}{Bare column name of location names to join; defaults \code{name}.}

\item{value}{Bare column name of numeric values to map; defaults \code{value}.}

\item{level}{String giving the desired geographic level; must be one of \code{"town"}, \code{"neighborhood"}, or \code{"tract"}. Defaults \code{"town"}.}

\item{city}{If geographic level is neighborhood, string of the corresponding city name to match to a spatial object.}

\item{n}{Number of breaks into which to bin values; defaults (approximately) 5.}

\item{palette}{String of a ColorBrewer palette; see \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::RColorBrewer()}} for possible values. Defaults \code{"GnBu"}.}

\item{title}{String giving the title, if desired, for the plot.}

\item{...}{Any other parameters to pass along to \code{geom_sf()}, such as \code{color} or \code{size}.}
}
\value{
A ggplot
}
\description{
This is a quick way to create a choropleth sketch of town-, neighborhood-, or tract-level data. Uses a corresponding \code{sf} object; as of June 2018, this \code{sf} object must be one that ships with this package, or otherwise be globally available and have a column \code{name}.
}
\examples{
\dontrun{
tidycensus::get_acs(geography = "county subdivision",
       variables = c(median_age = "B01002_001"), state = "09", county = "09") \%>\%
  town_names(NAME) \%>\%
  dplyr::filter(NAME \%in\% regions$`Greater New Haven`) \%>\%
  acs_quick_map(name = NAME, value = estimate, title = "Median age by town, 2017", n = 6)
}
}
