% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acs_vars18}
\alias{acs_vars18}
\title{Variable labels from the 2018 ACS}
\format{A data frame with 26996 rows and 3 variables:
\describe{
\item{name}{Variable code, where first 6 characters are the table number and last 3 digits are the variable number}
\item{label}{Readable label of each variable}
\item{concept}{Table name}
}}
\source{
US Census Bureau via \code{tidycensus}
}
\usage{
acs_vars18
}
\description{
Dataset of ACS variable labels, loaded from \code{tidycensus::load_variables()} for 2018 and cleaned up slightly.
}
\details{
This dataset is updated and renamed accordingly with each annual ACS release.
}
\keyword{datasets}
