% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acs_vars21}
\alias{acs_vars21}
\title{Variable labels from the 2021 ACS}
\format{
A data frame with 27886 rows and 3 variables:
\describe{
\item{name}{Variable code, where first 6 characters are the table number and last 3 digits are the variable number}
\item{label}{Readable label of each variable}
\item{concept}{Table name}
}
}
\source{
US Census Bureau via \code{tidycensus}
}
\usage{
acs_vars21
}
\description{
Dataset of ACS variable labels, loaded from \code{tidycensus::load_variables()} for 2021 and cleaned up slightly.
}
\details{
This dataset is updated and renamed accordingly with each annual ACS release.
}
\keyword{datasets}
