% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_logo.R
\name{add_logo}
\alias{add_logo}
\title{Easily add a logo to a ggplot}
\usage{
add_logo(plot, image = NULL, position = c("left", "right"), height = 0.05, ...)
}
\arguments{
\item{plot}{A \code{ggplot} object onto which the logo will be placed.}

\item{image}{Either a string giving the path or URL to an image file to be
read by \code{magick::image_read}; the results of reading a file or manipulating
an image already with \code{magick::image_read} or other \code{magick} functions;
a \code{ggplot} object / grob; some other object that can be handled by
\code{cowplot::draw_image}; or \code{NULL}, the default. If \code{NULL}, the image will
come from the file at \code{system.file("inst/extdata/logo.svg", package = "cwi")}.
As built, this is a logo for DataHaven, but that file can be replaced for
repackaging this library for other organizations or projects.}

\item{position}{String, either "left" or "right", giving the side on which
the logo should be aligned. Default: "left"}

\item{height}{Numeric: the height of the logo, as a percentage of the height
of the image given in \code{plot}. Adjust as necessary based on the dimensions
of the logo. Default: 0.05}

\item{...}{Additional arguments passed to \code{cowplot::draw_grob} if attaching
a grob, or to \code{cowplot::draw_image} otherwise.}
}
\value{
A \code{ggplot} object.
}
\description{
This function wraps around a few functions from \code{cowplot} and
\code{magick} to add a logo (or other annotation) to the bottom of a \code{ggplot} plot,
an otherwise tedious and easy to forget process. It's meant to be flexible in
the types of objects it can place; as a result, it's less flexible in their
placement and customization. For more specific needs, the source of this
function should be easy to build upon.
}
\examples{
if(interactive()){
  p <- ggplot2::ggplot(iris, ggplot2::aes(x = Sepal.Length)) +
     ggplot2::geom_density() +
       ggplot2::labs(title = "Test chart", caption = "Source: 2019 ACS 5-year estimates")

     add_logo(p)
     add_logo(p,
              magick::image_read("inst/extdata/25th_logo.png"), height = 0.1)

     # This example logo is not all that attractive, but shows how you might
     # attach a ggplot grob as a dynamically-created logo
     dummy_data <- data.frame(town = letters[1:4],
                              pop = c(21000, 40000, 81000, 36000))

     gg_logo <- ggplot2::ggplot(dummy_data, ggplot2::aes(x = town, y = pop)) +
       ggplot2::geom_col(width = 0.8, fill = "skyblue") +
       ggplot2::annotate(geom = "text", label = "DataHaven", x = 0.6, y = 6e4, hjust = 0,
                         family = "mono", size = 5) +
       ggplot2::theme_void()

     add_logo(p, gg_logo, width = 0.2, height = 0.1)
  }
}
\seealso{
\code{\link[magick:editing]{magick::image_read()}}, \code{\link[cowplot:draw_image]{cowplot::draw_image()}}
}
