% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basic_table_nums}
\alias{basic_table_nums}
\title{Basic ACS table numbers}
\format{
A named list of length 11, where names correspond to abbreviated subjects and string values correspond to table numbers.
}
\usage{
basic_table_nums
}
\description{
List of ACS table numbers commonly used by DataHaven for making short profiles of towns, neighborhoods, and regions.
}
\seealso{
\link{ext_table_nums}
}
\keyword{datasets}
