% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_csv_dump.R
\name{batch_csv_dump}
\alias{batch_csv_dump}
\title{Write a list of data frames to CSV and keep it movin'}
\usage{
batch_csv_dump(
  data,
  split_by,
  path = ".",
  base_name = NULL,
  bind = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{data}{A data frame or a list of data frames}

\item{split_by}{Bare column name of variable to split by. If \code{data} is a list, this is unnecessary and will be ignored.}

\item{path}{String giving a path at which to save files; defaults to current working directory.}

\item{base_name}{Optional string to be prepended to all file names.}

\item{bind}{Logical: whether to row-bind list of data frames into a single data frame. Defaults \code{FALSE}, in which case a list of data frames is returned.}

\item{verbose}{Logical: whether to print files' paths and names as they're written. Defaults \code{TRUE}.}

\item{...}{Additional arguments to pass to \code{utils::write.csv}, such as \code{na} or \code{col.names}.}
}
\value{
Either a list of data frames (in case of \code{bind = FALSE}) or a single data frame (in case of \code{bind = TRUE}).
}
\description{
This function takes either a list of data frames, or a data frame and a column to split by, and writes them all to CSV files. It then returns the list of data frames, optionally row-binded back together. It fits neatly in the middle of a longer piped workflow.
}
