% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cb_avail}
\alias{cb_avail}
\title{Tables available from the Census Bureau API}
\format{
A data frame with 48 rows and 4 variables:
\describe{
\item{vintage}{Vintage year of dataset}
\item{program}{Program, either "acs" or "dec"}
\item{survey}{Survey/summary file}
\item{title}{Brief title, standardized}
}
}
\source{
\href{https://www.census.gov/data/developers/updates/new-discovery-tool.html}{US Census Bureau API Discovery Tool}
}
\usage{
cb_avail
}
\description{
Dataset of tables available from the Census Bureau's API, filtered to include just mainland decennial summary files and ACS detailed tables. Useful for verifying which summary files are available for different years, as some codes have changed (3-year ACS was sunset, decennial summary files are now all demographic & housing characteristics tables). Also used internally for checking calls to \code{multi_geo_acs} and \code{multi_geo_decennial}.
}
\keyword{datasets}
