% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_n_wt.R
\name{collapse_n_wt}
\alias{collapse_n_wt}
\title{Collapse survey groups and get weighted means}
\usage{
collapse_n_wt(
  data,
  ...,
  .lvls,
  .group = group,
  .value = value,
  .weight = weight,
  .fill_wts = FALSE,
  .digits = NULL
)
}
\arguments{
\item{data}{A data frame, such as returned by \code{\link[dcws:xtab2df]{dcws::xtab2df()}} joined with
survey weights as returned by \code{\link[dcws:read_xtabs]{dcws::read_weights()}}. The default
column names here match those returned by \code{xtab2df} (\code{group}, \code{value}) and
\code{read_weights} (\code{weight}).}

\item{...}{Bare column names to use for grouping, including the \code{.group} column,
such as location, year, category, response, etc--probably everything except
values and weights.}

\item{.lvls}{A named list, where values are character vectors of smaller
groups (e.g. \code{c("<$15K", "$15K-$30K")}) and names are the groups those will
be replaced by (e.g. \code{"<$30K"}). This will be split into the arguments to
\code{\link[forcats:fct_collapse]{forcats::fct_collapse()}}.}

\item{.group}{Bare column name of where groups should be found. Default: group}

\item{.value}{Bare column name of where values should be found. Default: value}

\item{.weight}{Bare column name of where group weights should be found. Default: weight}

\item{.fill_wts}{Logical: if \code{TRUE}, missing weights will be filled in with 1,
i.e. unweighted. This defaults to \code{FALSE}, because missing weights is a
useful way to find that there's a mismatch between the group labels in
the data and those in the weights table, which is very often the case.
Therefore, only set this to \code{TRUE} if you've already accounted for labeling
discrepancies.}

\item{.digits}{Numeric: if given, weighted means will be rounded to this
number of digits. If \code{NULL} (the default), values are returned unrounded.}
}
\value{
A data frame with summarized values. The \code{.group} column will have
the collapsed groups, and the \code{.value} column will have average values.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \strong{Deprecation notice:} Crosstab-related
functions have been moved from cwi to the dcws package. The versions here will be
removed soon.
This is just a quick wrapper for a common, tedious task of
collapsing several demographic groups, such as income brackets, into
larger groups and taking a weighted mean based on a set of survey weights.
}
\seealso{
\code{\link[dcws:collapse_n_wt]{dcws::collapse_n_wt()}} \code{\link[dcws:xtab2df]{dcws::xtab2df()}} \code{\link[dcws:read_xtabs]{dcws::read_weights()}} \code{\link[forcats:fct_collapse]{forcats::fct_collapse()}}
}
\keyword{dcws-migration}
\keyword{internal}
