% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwi-package.R
\docType{package}
\name{cwi-package}
\alias{cwi}
\alias{cwi-package}
\title{cwi: Prep for DataHaven's Community Wellbeing Index}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This is a collection of functions for collecting and analyzing Census data and other sources to write the 2019 Community Wellbeing Index reports. It also contains reference tables for local geographies—neighborhoods, regions, etc—and commonly used ACS table numbers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CT-Data-Haven/cwi}
  \item \url{https://CT-Data-Haven.github.io/cwi}
  \item Report bugs at \url{https://github.com/CT-Data-Haven/cwi/issues}
}

}
\author{
\strong{Maintainer}: Camille Seaberry \email{camille@ctdatahaven.org} (\href{https://orcid.org/0000-0002-5102-995X}{ORCID})

Authors:
\itemize{
  \item Camille Seaberry \email{camille@ctdatahaven.org} (\href{https://orcid.org/0000-0002-5102-995X}{ORCID})
}

}
\keyword{internal}
