% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cws_demo}
\alias{cws_demo}
\title{CWS demo data}
\format{
A data frame with 100 rows and 7 variables:
\describe{
\item{code}{Question code}
\item{question}{Question text}
\item{category}{Category: gender, age, etc.}
\item{group}{Group: male, female, ages 18–34, etc.}
\item{response}{Survey response}
\item{value}{Percentage value}
\item{weight}{Survey weight}
}
}
\usage{
cws_demo
}
\description{
This is a sample of 2015 DataHaven Community Wellbeing Survey data for Greater New Haven with weights attached, saved here for use in examples.
}
\seealso{
Other example-datasets: 
\code{\link{gnh_tenure}}
}
\concept{example-datasets}
\keyword{datasets}
