% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dh_scaffold.R
\name{dh_scaffold}
\alias{dh_scaffold}
\title{dh_scaffold}
\usage{
dh_scaffold(
  dir = ".",
  input_data = TRUE,
  output_data = TRUE,
  fetch_data = TRUE,
  analysis = TRUE,
  prep_scripts = FALSE,
  plots = FALSE,
  format_tables = FALSE,
  drafts = FALSE,
  utils = TRUE,
  addl = NULL,
  gitblank = TRUE
)
}
\arguments{
\item{dir}{String: path to directory in which new files will be written. Default: '.'}

\item{input_data}{Create a directory \code{input_data}. Default: TRUE.

Standard use: data from some outside source to be analyzed in this project.}

\item{output_data}{Create a directory \code{output_data}. Default: TRUE.

Standard use: data written after analysis done in this project, generally in formats that can still be used for analysis and visualization (csv, rds) rather than formats for distribution (I usually add a folder \code{to_distro}) or to pass on to a client (xlsx). Nice spreadsheet outputs should go in \code{format_tables} or some other distribution-centered folder.}

\item{fetch_data}{Create a directory \code{fetch_data}. Default: TRUE.

Standard use: a place to dump data as it comes in from API calls, queries, batch file downloads, etc.}

\item{analysis}{Create a directory \code{analysis}. Default: TRUE

Standard use: main analysis scripts, both notebooks and .R scripts.}

\item{prep_scripts}{Create a directory \code{prep_scripts}. Default: FALSE

Standard use: scripts use to prep or reshape data or documents, e.g. creating formatted spreadsheets for a client, making metadata, prepping to post to data.world, formatting for a website, bulk rendering parameterized Rmarkdown documents.}

\item{plots}{Create a directory \code{plots}. Default: FALSE

Standard use: plots, either for in-house use or outside distribution.}

\item{format_tables}{Create a directory \code{format_tables}. Default: FALSE

Standard use: spreadsheets--probably written by a script in \code{prep_scripts}--to be shared with clients or collaborators. Think of these as being files appropriate for presentation or addenda to a report, not for doing further analysis.}

\item{drafts}{Create a directory \code{drafts}. Default: FALSE

Standard use: separating the more EDA-centered notebooks from notebooks used for drafting writing. Also a good place to keep files that have been edited in outside software (.docx, etc).}

\item{utils}{Create a directory \code{utils}. Default: TRUE

Standard use: utility scripts and miscellaneous files, e.g. logo images, snippets of data, lists of colors to use.}

\item{addl}{A string vector of any additional directories to create. Default: NULL}

\item{gitblank}{Logical: whether to write a blank placeholder file in each new directory to force git tracking, even without yet having folder contents. Default: TRUE. If FALSE, empty directories will \emph{not} be tracked by git.}
}
\value{
Returns nothing, but prints paths to newly created directories.
}
\description{
Setup directories for a typical DataHaven project
}
\details{
This sets up a typical project directory structure that we use for many projects at DataHaven. It will write directories at the specified path, but it will NOT overwrite any directories that already exist. You'll have the option to cancel before anything is written.
}
