% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{education}
\alias{education}
\title{ACS demo data - educational attainment}
\format{
A data frame with 125 rows and 4 variables:
\describe{
\item{name}{Town name}
\item{edu_level}{Educational attainment}
\item{estimate}{Estimated count}
\item{moe}{Margin of error}
}
}
\usage{
education
}
\description{
This is a table of educational attainment data for adults ages 25 and up in a few New Haven-area towns obtained with \code{tidycensus::get_acs}, to be used in testing and examples.
}
\keyword{example-datasets}
