% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ext_table_nums}
\alias{ext_table_nums}
\title{Extended ACS table numbers}
\format{
A named list of length 26, where names correspond to abbreviated subjects and string values correspond to table numbers.
}
\usage{
ext_table_nums
}
\description{
List of ACS tables less commonly used by DataHaven. These are used for "extended" profiles, such as neighborhood profiles distributed as PDF on the DataHaven website.
}
\seealso{
\link{basic_table_nums}
}
\keyword{datasets}
