% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gnh_tenure}
\alias{gnh_tenure}
\title{ACS demo data - tenure}
\format{
A data frame with 45 rows and 5 variables:
\describe{
\item{level}{Geographic level}
\item{name}{Geography name}
\item{tenure}{Tenure: total households, owner-occupied, or renter-occupied}
\item{estimate}{Estimated count}
\item{share}{Calculated share of households, or \code{NA} for total}
}
}
\usage{
gnh_tenure
}
\description{
This is a table of housing tenure data for Greater New Haven obtained with \code{multi_geo_acs}, to be used in testing and examples.
}
\keyword{datasets}
