% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_census.R
\name{label_decennial}
\alias{label_decennial}
\alias{label_acs}
\title{Quickly add the labels of census variables}
\usage{
label_decennial(data, year = 2020, sumfile = "dhc", variable = variable)

label_acs(data, year = 2023, survey = "acs5", variable = variable)
}
\arguments{
\item{data}{A data frame/tibble.}

\item{year}{The year of data; defaults to 2024 for ACS, or 2020 for decennial.}

\item{sumfile}{For \code{label_decennial}, a string: which summary file to use. Defaults to \code{"dhc"}, the code used for 2020. 2010 used summary files labeled \code{"sf1"} or \code{"sf3"}.}

\item{variable}{The bare column name of variable codes; defaults to \code{variable}, as returned by the \verb{multi_geo_*} or \verb{tidycensus::get_*} functions.}

\item{survey}{For \code{label_acs}, a string: which ACS estimate to use. Defaults to 5-year (\code{"acs5"}), but can also be 1-year (\code{"acs1"}) or 3-year (\code{"acs3"}), though both 1-year and 3-year have limited availability.}
}
\value{
A tibble with the same number of rows as \code{data} but an additional column called \code{label}
}
\description{
\verb{multi_geo_*} functions, and their underlying \verb{tidycensus::get_*} functions, return data tables with variable codes (e.g. "B01001_003"), which can be joined with lookup tables to get readable labels (e.g. "Total!!Male!!Under 5 years"). These functions are just quick wrappers around the common task of joining your data frame with the variable codes and labels.
}
\examples{
\dontrun{
acs_pops <- multi_geo_acs("B01001")
label_acs(acs_pops)
}
}
\seealso{
\link{decennial_vars} \link{acs_vars}
}
\keyword{augmenting-functions}
