% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_census.R
\name{label_decennial}
\alias{label_decennial}
\title{Quickly add the labels of decennial variables}
\usage{
label_decennial(data, year = 2020, sumfile = "dhc", variable = variable)
}
\arguments{
\item{data}{A data frame/tibble.}

\item{year}{The year of decennial census data; defaults 2020.}

\item{sumfile}{A string: which summary file to use. Defaults to \code{"dhc"}, the code used for 2020. 2010 used summary files labeled \code{"sf1"} or \code{"sf3"}.}

\item{variable}{The bare column name of variable codes; defaults to \code{variable}, as returned by \code{tidycensus::get_decennial}.}
}
\value{
A tibble
}
\description{
\code{tidycensus::get_decennial} returns a decennial data table with its variable codes, which can be joined with \code{cwi::decennial_vars20} to get readable labels. This function is just a quick wrapper around the common task of joining these two data frames.
}
\seealso{
\link{decennial_vars}
}
