% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{laus_codes}
\alias{laus_codes}
\title{LAUS area codes}
\format{
A data frame with 6625 rows and 4 variables:
\describe{
\item{type}{Area type code}
\item{state_code}{Two-digit state FIPS code}
\item{area}{Area name}
\item{area_code}{Area code}
}
}
\source{
This is a filtered and cleaned version of a file downloaded from \href{https://download.bls.gov/pub/time.series/la/la.area}{BLS}.
}
\usage{
laus_codes
}
\description{
A dataset of area types and codes for states, counties, and towns across the US, as used for the Local Area Unemployment Statistics. These are needed to put together series names, like those used to make API calls in \code{qwi_industry}. This used to be filtered just for Connecticut.
}
\keyword{datasets}
