% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laus_trend.R
\name{laus_trend}
\alias{laus_trend}
\title{Fetch local area unemployment statistics (LAUS) data over time}
\usage{
laus_trend(
  names = NULL,
  startyear,
  endyear,
  state = "09",
  measures = "all",
  annual = FALSE,
  verbose = TRUE,
  key = NULL
)
}
\arguments{
\item{names}{A character vector of place names to look up, either towns and/or counties.}

\item{startyear}{Numeric; first year of range}

\item{endyear}{Numeric; last year of range}

\item{state}{A string: either name or two-digit FIPS code of a US state. Required; defaults \code{"09"} (Connecticut).}

\item{measures}{A character vector of measures, containing any combination of \code{"unemployment rate"}, \code{"unemployment"}, \code{"employment"}, or \code{"labor force"}, or \code{"all"} (the default) as shorthand for all of the above.}

\item{annual}{Logical: whether to include annual averages along with monthly data. Defaults \code{FALSE}.}

\item{verbose}{Logical: if \code{TRUE} (default), this will print overview information about the series being used, as returned by the API.}

\item{key}{A string giving the BLS API key. If \code{NULL} (the default), will take the value in \code{Sys.getenv("BLS_KEY")}.}
}
\value{
A data frame, slightly cleaned up from what the API returns.
}
\description{
Fetch monthly LAUS data for a list of locations over a given time period, modeled after \code{blscrapeR::bls_api}. Requires a BLS API key.
}
\examples{
\dontrun{
laus_trend(c("Connecticut", "New Haven", "Hamden"), 2014, 2017, annual = TRUE)
}
}
\keyword{fetching-functions}
