% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grps.R
\name{make_grps}
\alias{make_grps}
\title{Collapse variable into subgroup positions, for use in \code{add_grps}.}
\usage{
make_grps(x, grps)
}
\arguments{
\item{x}{Character vector}

\item{grps}{Named list of variable values or positions within vector \code{x}}
}
\value{
Named list of variable values from given positions
}
\description{
Collapse variable into subgroup positions, for use in \code{add_grps}.
}
\keyword{internal}
