% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msa}
\alias{msa}
\title{Names and GEOIDs of regional MSAs}
\format{
A data frame with 392 rows and 3 variables:
\describe{
\item{geoid}{GEOID/FIPS code}
\item{name}{Name of MSA}
\item{region}{String: whether MSA is inside or outside of New England}
}
}
\source{
US Census Bureau via \code{tidycensus}. Note that these are as of 2020.
}
\usage{
msa
}
\description{
A reference dataset of all the metropolitan statistical areas (MSAs) in the US, marked with whether they're in a New England state.
}
\keyword{datasets}
