% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_geo_fetch.R
\name{multi_geo_decennial}
\alias{multi_geo_decennial}
\title{Fetch a decennial census table with multiple geography levels}
\usage{
multi_geo_decennial(
  table,
  year = endyears[["decennial"]],
  towns = "all",
  regions = NULL,
  counties = "all",
  state = "09",
  neighborhoods = NULL,
  tracts = NULL,
  blockgroups = NULL,
  msa = FALSE,
  us = FALSE,
  new_england = TRUE,
  nhood_name = "name",
  nhood_geoid = NULL,
  nhood_weight = "weight",
  sumfile = c("dhc", "sf1", "sf3", "pl"),
  verbose = TRUE,
  key = NULL,
  sleep = 0,
  ...
)
}
\arguments{
\item{table}{A string giving the decennial census table number. These are generally formatted as one or more letters, 3 numbers, and optionally a letter.}

\item{year}{The year of the census table; currently defaults 2020.}

\item{towns}{A character vector of towns to include; \code{"all"} (default) for all towns optionally filtered by county; or \code{NULL} to not fetch town-level table.}

\item{regions}{A named list of regions with their town names (defaults \code{NULL}).}

\item{counties}{A character vector of counties to include; \code{"all"} (default) for all counties in the state; or \code{NULL} to not fetch county-level table.}

\item{state}{A string: either name or two-digit FIPS code of a US state. Required; defaults \code{"09"} (Connecticut).}

\item{neighborhoods}{A data frame with columns for neighborhood name, GEOID of tracts, and weight, e.g. share of each tract assigned to a neighborhood. If included, weighted sums will be returned for neighborhoods. Unlike \code{multi_geo_acs}, this doesn't take block groups.}

\item{tracts}{A character vector of 11-digit FIPS codes of tracts to include, or \code{"all"} for all tracts optionally filtered by county. Defaults \code{NULL}.}

\item{blockgroups}{A character vector of 12-digit FIPS codes of block groups to include, or \code{"all"} for all block groups optionally filtered by county. Defaults \code{NULL}.}

\item{msa}{Logical: whether to fetch New England states' metropolitan statistical areas. Defaults \code{FALSE}.}

\item{us}{Logical: whether to fetch US-level table. Defaults \code{FALSE}.}

\item{new_england}{Logical: if \code{TRUE} (the default), limits metro areas to just New England states.}

\item{nhood_name}{String giving the name of the column in the data frame \code{neighborhoods} that contains neighborhood names. Previously this was a bare column name, but for consistency with changes to COG-based FIPS codes, this needs to be a string. Only relevant if a neighborhood weight table is being used. Defaults \code{"name"} to match the neighborhood lookup datasets.}

\item{nhood_geoid}{String giving the name of the column in \code{neighborhoods} that contains neighborhood GEOIDs, either tracts or block groups. Only relevant if a neighborhood weight table is being used. Because of changes to FIPS codes, this no longer has a default.}

\item{nhood_weight}{String giving the name of the column in \code{neighborhoods} that contains weights between neighborhood names and tract/block groups. Only relevant if a neighborhood weight table is being used. Defaults \code{"weight"} to match the neighborhood lookup datasets.}

\item{sumfile}{A string giving the summary file to pull from. Note that codes have changed between 2010 and 2020. Now that default year is 2020, default sumfile is \code{"dhc"}. For 2010, should be either \code{"sf1"}, or less commonly \code{"sf3"}. Use \code{"pl"} for 2020 redistricting data.}

\item{verbose}{Logical: whether to print summary of geographies included. Defaults \code{TRUE}.}

\item{key}{String: Census API key. If \code{NULL} (default), takes the value from \code{Sys.getenv("CENSUS_API_KEY")}.}

\item{sleep}{Number of seconds, if any, to sleep before each API call. This might help with the Census API's tendency to crash, but for many geographies, it could add a sizable about of time. Probably don't add more than a few seconds.}

\item{...}{
  Arguments passed on to \code{\link[tidycensus:get_decennial]{tidycensus::get_decennial}}
  \describe{
    \item{\code{variables}}{Character string or vector of character strings of variable
IDs.}
    \item{\code{county}}{The county for which you are requesting data. County names and
FIPS codes are accepted. Must be combined with a value supplied
to `state`.  Defaults to NULL.}
    \item{\code{geometry}}{if FALSE (the default), return a regular tibble of ACS data.
if TRUE, uses the tigris package to return an sf tibble
with simple feature geometry in the `geometry` column.}
    \item{\code{output}}{One of "tidy" (the default) in which each row represents an
enumeration unit-variable combination, or "wide" in which each
row represents an enumeration unit and the variables are in the
columns.}
    \item{\code{keep_geo_vars}}{if TRUE, keeps all the variables from the Census
shapefile obtained by tigris.  Defaults to FALSE.}
    \item{\code{shift_geo}}{(deprecated) if TRUE, returns geometry with Alaska and Hawaii
shifted for thematic mapping of the entire US.
Geometry was originally obtained from the albersusa R package.  As of May 2021,
we recommend using \code{tigris::shift_geometry()} instead.}
    \item{\code{summary_var}}{Character string of a "summary variable" from the decennial Census
to be included in your output. Usually a variable (e.g. total population)
that you'll want to use as a denominator or comparison.}
    \item{\code{pop_group}}{The population group code for which you'd like to request data.  Applies to summary files for which population group breakdowns are available like the Detailed DHC-A file.}
    \item{\code{pop_group_label}}{If \code{TRUE}, return a \code{"pop_group_label"} column that contains the label for the population group.  Defaults to \code{FALSE}.}
    \item{\code{show_call}}{if TRUE, display call made to Census API. This can be very useful
in debugging and determining if error messages returned are
due to tidycensus or the Census API. Copy to the API call into
a browser and see what is returned by the API directly. Defaults to FALSE.}
  }}
}
\value{
A tibble with GEOID, name, variable code, estimate, moe, geography level, state, and year, as applicable, for the chosen table.
}
\description{
Fetch a data table from the decennial census via \code{tidycensus} with your choice of geographies at multiple levels. For geographies made of aggregates, i.e. neighborhoods made of tracts or regions made of towns, the returned table will have estimates summed for the whole area. Any geographic levels that are null will be excluded.
}
\details{
This function essentially calls \code{\link[tidycensus:get_decennial]{tidycensus::get_decennial()}} multiple times, depending on geographic levels chosen, and does minor cleaning, filtering, and aggregation. Note that the underlying \code{\link[tidycensus:get_decennial]{tidycensus::get_decennial()}} requires a Census API key. As is the case with other \code{tidycensus} functions, \code{multi_geo_decennial} assumes this key is stored as \code{CENSUS_API_KEY} in your \code{.Renviron} or other source of environment variables. See \code{\link[tidycensus:census_api_key]{tidycensus::census_api_key()}} for installation.

Be advised that decennial table numbers generally change from year to year, so if you're looking at trends, check \href{https://data.census.gov}{data.census.gov} or another source to make sure the tables have the same meaning. Setting \code{verbose = TRUE} is helpful for this as well.
}
\examples{
\dontrun{
multi_geo_decennial("P1", 2020,
    sumfile = "dhc",
    towns = "all",
    regions = list(inner_ring = c("Hamden", "East Haven", "West Haven")),
    counties = "New Haven County"
)
}
}
\seealso{
\code{\link[tidycensus:census_api_key]{tidycensus::census_api_key()}}, \code{\link[tidycensus:get_decennial]{tidycensus::get_decennial()}}
}
\keyword{fetching-functions}
