% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sf.R
\docType{data}
\name{neighborhood_shapes}
\alias{neighborhood_shapes}
\alias{bridgeport_sf}
\alias{hartford_sf}
\alias{new_haven_sf}
\alias{stamford_sf}
\title{Neighborhood shapefiles of Connecticut cities}
\format{
Data frames (classes \code{data.frame} and \code{sf}). In each, \code{name} is the name of each neighborhood, and \code{geometry} is the shape of each neighborhood as either a \code{sfc_POLYGON} or \code{sfc_MULTIPOLYGON} object. \code{hartford_sf} contains an additional variable, \code{town}, which marks the neighborhoods as being in Hartford or West Hartford.

An object of class \code{sf} (inherits from \code{data.frame}) with 13 rows and 2 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 27 rows and 3 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 20 rows and 2 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 13 rows and 2 columns.
}
\source{
All neighborhood boundaries, with the exception of West Hartford, come directly from their respective cities' online data portals. West Hartford neighborhood boundaries are based directly on current tracts; the common names corresponding to tracts were scraped from a third-party real estate site.
}
\usage{
bridgeport_sf

hartford_sf

new_haven_sf

stamford_sf
}
\description{
Data frames of neighborhoods of Connecticut cities with \code{sf} geometries. Includes Bridgeport, Hartford/West Hartford, New Haven, and Stamford.
}
\keyword{datasets}
\keyword{shapefiles}
