% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhv_bgrps}
\alias{nhv_bgrps}
\title{New Haven neighborhoods by block group}
\format{A data frame with 110 rows and 5 variables:
\describe{
\item{neighborhood}{Neighborhood name}
\item{tract}{6-digit FIPS code of the tract}
\item{block_group}{Single digit number of the block group}
\item{weight}{Share of block group population in that neighborhood}
\item{geoid}{12-digit FIPS code of the block group}
}}
\usage{
nhv_bgrps
}
\description{
A dataset of New Haven neighborhoods delineated by Census tract and block group. Some block groups cross between more than one neighborhood; use \code{weight} column for aggregating values such as populations.
}
\keyword{datasets}
