% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhv_tracts}
\alias{nhv_tracts}
\title{New Haven neighborhoods by tract}
\format{A data frame with 34 rows and 4 variables:
\describe{
\item{neighborhood}{Neighborhood name}
\item{tract}{6-digit FIPS code of the tract}
\item{weight}{Share of block group population in that neighborhood}
\item{geoid}{Full 11-digit FIPS code of the tract}
}}
\usage{
nhv_tracts
}
\description{
A dataset of New Haven neighborhoods delineated by Census tract. Some tracts cross between more than one neighborhood; use \code{weight} column for aggregating values such as populations.
}
\keyword{datasets}
