% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{occ_codes}
\alias{occ_codes}
\title{Census occupation codes}
\format{
A data frame with 32 rows and 5 columns:
\describe{
\item{is_major_grp}{Logical: whether this is the highest level included}
\item{occ_group}{Major occupation group name}
\item{occ_code}{Census occupation code}
\item{soc_code}{SOC code}
\item{description}{Full text of occupation name}
}
}
\source{
US Census Bureau's industry & occupation downloads
}
\usage{
occ_codes
}
\description{
A dataset of occupation groups and descriptions with both Census (OCC) codes and SOC codes. Occupations are grouped hierarchically. This is filtered from a Census crosswalk to include only top-level groups, except for the very broad management, business, science, and arts occupations group; for this one, the second level groups are treated as the major one. Often you'll just want the major groups, so you can filter by the \code{is_major_grp} column.
}
\keyword{datasets}
