% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regions}
\alias{regions}
\title{Regions of Connecticut}
\format{
A named list of vectors, where names give the names of regions and vectors give the names of towns making up each region, including regional councils of governments.
}
\source{
DataHaven internal and CT OPM
}
\usage{
regions
}
\description{
A dataset of Connecticut regions by town
}
\seealso{
Other ref-datasets: 
\code{\link{acs_vars}},
\code{\link{ct5_clusters}},
\code{\link{decennial_vars}},
\code{\link{msa}},
\code{\link{naics_codes}},
\code{\link{neighborhood_tracts}},
\code{\link{occ_codes}},
\code{\link{proxy_pumas}},
\code{\link{school_dists}},
\code{\link{table_nums}},
\code{\link{village2town}},
\code{\link{xwalk}},
\code{\link{zip2town}}
}
\concept{ref-datasets}
\keyword{datasets}
