% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availability.R
\name{safe_read_avail}
\alias{safe_read_avail}
\title{Check availability of different APIs with error catching}
\usage{
safe_read_avail(url, type, query = NULL)
}
\arguments{
\item{url}{URL for API call}

\item{type}{File type, currently json or html}

\item{query}{Optional query args as a list, added in order to pass a key to census api}
}
\description{
Check availability of different APIs with error catching
}
\keyword{internal}
