% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{school_dists}
\alias{school_dists}
\title{School districts by town}
\format{
A data frame with 196 rows and 2 variables:
\describe{
\item{district}{School district name}
\item{town}{Name of town included in district}
}
}
\source{
Distinct town-level district names come from the state's \href{https://data.world/state-of-connecticut/9k2y-kqxn}{data.world}. Regional districts and their towns come from combing through school district websites.
}
\usage{
school_dists
}
\description{
A crosswalk of Connecticut towns and school districts, including regional districts. Some very small towns don't operate their own schools and are not included here, whereas other towns are both part of a regional district and operate some of their own schools, most commonly their own elementary.
}
\details{
Note that for aggregating by region, the Hartford-area regional district CREC is not included here, because it spans so many other towns that also run their own schools.CREC is of significant enough size that it should generally be included in Greater Hartford aggregations or as its own district for comparisons alongside e.g. Hartford and West Hartford.
}
\keyword{datasets}
