% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_acs.R
\name{separate_acs}
\alias{separate_acs}
\title{Separate labels given to ACS data}
\usage{
separate_acs(
  data,
  col = label,
  into = NULL,
  sep = "!!",
  drop_total = FALSE,
  ...
)
}
\arguments{
\item{data}{A data frame such as returned by \code{\link[=multi_geo_acs]{multi_geo_acs()}} or \code{\link[tidycensus:get_acs]{tidycensus::get_acs()}}.}

\item{col}{Bare column name where ACS labels are. Default: label}

\item{into}{Character vector of names of new variables. If \code{NULL} (the default), names will be assigned as "x1", "x2," etc. If you don't want to include the Total column, this character vector only needs to include the groups other than Total (see examples).}

\item{sep}{Character: separator between columns. Default: '!!'}

\item{drop_total}{Logical, whether to include the "Total" column that comes from separating ACS data. Default: FALSE}

\item{...}{
  Arguments passed on to \code{\link[tidyr:separate]{tidyr::separate}}
  \describe{
    \item{\code{remove}}{If \code{TRUE}, remove input column from output data frame.}
    \item{\code{convert}}{If \code{TRUE}, will run \code{\link[=type.convert]{type.convert()}} with
\code{as.is = TRUE} on new columns. This is useful if the component
columns are integer, numeric or logical.

NB: this will cause string \code{"NA"}s to be converted to \code{NA}s.}
    \item{\code{extra}}{If \code{sep} is a character vector, this controls what
happens when there are too many pieces. There are three valid options:
\itemize{
\item \code{"warn"} (the default): emit a warning and drop extra values.
\item \code{"drop"}: drop any extra values without a warning.
\item \code{"merge"}: only splits at most \code{length(into)} times
}}
    \item{\code{fill}}{If \code{sep} is a character vector, this controls what
happens when there are not enough pieces. There are three valid options:
\itemize{
\item \code{"warn"} (the default): emit a warning and fill from the right
\item \code{"right"}: fill with missing values on the right
\item \code{"left"}: fill with missing values on the left
}}
  }}
}
\value{
A data frame
}
\description{
This is a quick wrapper around \code{\link[tidyr:separate]{tidyr::separate()}} written to match the standard formatting used for ACS variable labels. These generally take the form e.g. "Total!!Male!!5 to 9 years". This function will separate values by \code{"!!"} and optionally drop the resulting "Total" column, which is generally constant for the entire data frame.
}
\examples{
\dontrun{
if (interactive()) {
    age <- label_acs(multi_geo_acs("B01001"))

    # Default: allow automatic labeling, in this case x1, x2, x3
    separate_acs(age)

    # Drop Total column, use automatic labeling (x1 & x2)
    separate_acs(age, drop_total = TRUE)

    # Keep Total column; assign names total, sex, age
    separate_acs(age, into = c("total", "sex", "age"))

    # Drop Total column; only need to name sex & age
    separate_acs(age, into = c("sex", "age"), drop_total = TRUE)

    # Carried over from tidyr::separate, using NA in place of the Total column
    # will also drop that column and yield the same as the previous example
    separate_acs(age, into = c(NA, "sex", "age"))
}
}
}
\seealso{
\code{\link[tidyr:separate]{tidyr::separate()}}
}
