% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_uniq.R
\name{show_uniq}
\alias{show_uniq}
\title{Print unique values from a data frame column, then keep it moving}
\usage{
show_uniq(data, col)
}
\arguments{
\item{data}{A data frame}

\item{col}{Bare column name of interest}
}
\value{
Original unchanged \code{data}
}
\description{
\code{show_uniq} gets the unique values of a column and their position within that vector, prints them neatly to the console, then returns the original data frame unchanged. It's just a convenience for showing the values in a column without breaking your workflow or train of thought, and is useful for identifying groups for \code{add_grps}.
}
\examples{
# show_uniq makes it easy to see that the values of `edu_level` that correspond
# to less than high school are in positions 2-16, and so on
education |>
    dplyr::group_by(name) |>
    show_uniq(edu_level) |>
    add_grps(
        list(
            ages25plus = 1,
            less_than_high_school = 2:16,
            high_school_plus = 17:25,
            bachelors_plus = 22:25
        ),
        group = edu_level, value = estimate
    )
}
\keyword{utils}
