% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_nums}
\alias{table_nums}
\alias{basic_table_nums}
\alias{ext_table_nums}
\title{Common ACS table numbers}
\format{
A named list, where names correspond to abbreviated subjects and string values correspond to table numbers.

An object of class \code{list} of length 11.

An object of class \code{list} of length 26.
}
\usage{
basic_table_nums

ext_table_nums
}
\description{
Lists of ACS table numbers commonly used by DataHaven. \code{basic_table_nums} is used for making short profiles of towns, neighborhoods, and regions, while \code{ext_table_nums} is used for "extended" profiles, such as neighborhood profiles distributed as PDF on the DataHaven website.
}
\keyword{datasets}
