% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/town_names.R
\name{town_names}
\alias{town_names}
\title{Clean up town names as returned from ACS}
\usage{
town_names(.data, name_col)
}
\arguments{
\item{.data}{A data frame}

\item{name_col}{Bare column name of town names}
}
\value{
A tibble/data frame with cleaned names and "not defined" towns removed
}
\description{
Clean up town names as returned from ACS
}
