% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sf.R
\docType{data}
\name{town_sf}
\alias{town_sf}
\title{Connecticut town shapefile}
\format{
Data frame of class \code{sf}.
\describe{
\item{name}{Town name}
\item{GEOID}{Town GEOID}
\item{geometry}{Town geometry as \code{sfc_MULTIPOLYGON}}
}
}
\usage{
town_sf
}
\description{
Data frame of Connecticut towns with \code{sf} geometries.
}
\keyword{datasets}
