% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sf.R
\docType{data}
\name{tract_sf}
\alias{tract_sf}
\alias{tract_sf19}
\title{Connecticut tract shapefile}
\format{
Data frame of class \code{sf}.
\describe{
\item{name}{Tract FIPS code}
\item{geometry}{Tract geometry as \code{sfc_MULTIPOLYGON}}
}

An object of class \code{sf} (inherits from \code{data.frame}) with 829 rows and 2 columns.
}
\usage{
tract_sf

tract_sf19
}
\description{
Data frame of Connecticut tracts with \code{sf} geometries. \code{tract_sf19} is the 2019 version, should you need an older shapefile.
}
\keyword{datasets}
\keyword{shapefiles}
