% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{village2town}
\alias{village2town}
\title{Village to town lookup}
\format{
A data frame with 90 rows and 4 variables:
\describe{
\item{cdp_geoid}{7-digits CDP FIPS code}
\item{place}{CDP name}
\item{town_geoid}{10-digit town FIPS code}
\item{town}{Town name}
}
}
\source{
Spatial overlay of TIGER shapefiles, plus manual additions of non-CDP villages when we run into them in the wild.
}
\usage{
village2town
}
\description{
A crosswalk of Connecticut's Census-designated places (CDP) to the towns (county subdivisions) containing them.
}
\keyword{datasets}
