% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xwalk}
\alias{xwalk}
\alias{tract2town}
\title{CT crosswalk}
\format{
\subsection{For \code{xwalk}:}{

A data frame with 48358 rows and 17 variables:
\describe{
\item{block}{Block FIPS code}
\item{block_grp}{Block group FIPS code, based on county FIPS}
\item{block_grp_cog}{Block group FIPS code, based on COG FIPS as of 2022 ACS}
\item{tract}{Tract FIPS code, based on county FIPS}
\item{tract_cog}{Tract FIPS code, based on COG FIPS as of 2022 ACS}
\item{town}{Town name}
\item{town_fips}{Town FIPS code, based on county FIPS}
\item{town_fips_cog}{Town FIPS code, based on COG FIPS as of 2022 ACS}
\item{county}{County name}
\item{county_fips}{County FIPS code}
\item{cog}{COG name}
\item{cog_fips}{COG FIPS code}
\item{msa}{Metro/micropolitan area name}
\item{msa_fips}{Metro/micropolitan area FIPS code}
\item{puma}{PUMA name}
\item{puma_fips}{PUMA FIPS code, based on county FIPS}
\item{puma_fips_cog}{PUMA FIPS code, based on COG FIPS as of 2022 ACS}
}
}

\subsection{For \code{tract2town}:}{

A data frame with 879 rows and 3 variables:
\describe{
\item{tract}{Tract FIPS code}
\item{tract_cog}{Tract FIPS code, based on COG FIPS as of 2022 ACS}
\item{town}{Town name}
}
}
}
\source{
2020 and 2022 (for COGs & towns) TIGER shapefiles
}
\usage{
xwalk

tract2town
}
\description{
A crosswalk between geographies in Connecticut, built off of TIGER shapefiles. \code{tract2town} is a subset with just tracts and towns. Tracts that span multiple towns are deduplicated and listed with the town with the largest areal overlap.

A version of \code{cwi::xwalk} that only contains tracts and town names, and is deduplicated for tracts that span multiple towns.
}
\seealso{
Other ref-datasets: 
\code{\link{acs_vars}},
\code{\link{ct5_clusters}},
\code{\link{decennial_vars}},
\code{\link{msa}},
\code{\link{naics_codes}},
\code{\link{neighborhood_tracts}},
\code{\link{occ_codes}},
\code{\link{proxy_pumas}},
\code{\link{regions}},
\code{\link{school_dists}},
\code{\link{table_nums}},
\code{\link{village2town}},
\code{\link{zip2town}}

Other ref-datasets: 
\code{\link{acs_vars}},
\code{\link{ct5_clusters}},
\code{\link{decennial_vars}},
\code{\link{msa}},
\code{\link{naics_codes}},
\code{\link{neighborhood_tracts}},
\code{\link{occ_codes}},
\code{\link{proxy_pumas}},
\code{\link{regions}},
\code{\link{school_dists}},
\code{\link{table_nums}},
\code{\link{village2town}},
\code{\link{zip2town}}
}
\concept{ref-datasets}
\keyword{datasets}
