% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zip2town}
\alias{zip2town}
\title{Zip to town lookup}
\format{
A data frame with 407 rows and 7 variables:
\describe{
\item{zip}{5-digit zip code (ZCTA5)}
\item{town_geoid}{10-digit town FIPS code}
\item{town}{Town name}
\item{poppt}{Population living in this combination of zip and town}
\item{hupt}{Number of housing units in this combination of zip and town}
\item{pct_zip_hh_in_town}{Percentage of the zip's total households that are also in this town}
\item{pct_town_hh_in_zip}{Percentage of the town's total households that are also in this zip}
}
}
\source{
Cleaned-up version of the Census \href{https://www.census.gov/geo/maps-data/data/zcta_rel_download.html}{2010 ZCTA to county subdivision relationship file}
}
\usage{
zip2town
}
\description{
A crosswalk of Connecticut's ZCTA5s, towns, and town GEOIDs with shares of households in overlapping areas. Each row corresponds to a combination of zip and town; therefore, some zips have more than one observation, as do towns.
}
\keyword{datasets}
