% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acs_vars}
\alias{acs_vars}
\title{Variable labels from the ACS}
\format{A data frame with 22815 rows and 3 variables:
\describe{
\item{name}{Variable code, where first 6 characters are the table number and last 3 digits are the variable number}
\item{label}{Readable label of each variable}
\item{concept}{Table name}
}}
\source{
US Census Bureau via \code{tidycensus}
}
\usage{
acs_vars
}
\description{
Dataset of ACS variable labels, loaded from \code{tidycensus::load_variables()} and cleaned up slightly.
}
\keyword{datasets}
