% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acs_vars}
\alias{acs_vars}
\title{Variable labels from the 2023 ACS}
\format{
A data frame with 28261 rows and 3 variables:
\describe{
\item{name}{Variable code, where first 6 characters are the table number and last 3 digits are the variable number}
\item{label}{Readable label of each variable}
\item{concept}{Table name}
}
}
\source{
US Census Bureau via \code{tidycensus}
}
\usage{
acs_vars
}
\description{
Dataset of ACS variable labels, loaded from \code{tidycensus::load_variables()} for 2023 and cleaned up slightly.
}
\details{
This dataset is updated with each annual ACS release, with an attribute \code{year} giving the ACS endyear of the dataset.
}
\examples{
# get the year
attr(acs_vars, "year")
}
\keyword{datasets}
