% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availability.R
\name{check_cb_avail}
\alias{check_cb_avail}
\alias{check_qwi_avail}
\title{Check availability of datasets}
\usage{
check_cb_avail()

check_qwi_avail()
}
\value{
\strong{For \code{check_cb_avail}}: A data frame with columns for vintage, program (e.g. "acs"), survey (e.g. "acs5"),
and title, as returned from the Census Bureau API.

\strong{For check_qwi_avail`}: A data frame with columns for state FIPS code, earliest year available, and most recent year available.
}
\description{
These two functions check for the availability of datasets needed
to formulate API queries.
\itemize{
\item \code{check_cb_avail} finds all available vintages of
major surveys under the Census Bureau's ACS and Decennial programs for the
mainland US.
\item \code{check_qwi_avail} finds all years of QWI data available per state.
}

Previously, these were datasets built into the package, which ran the risk of
being outdated and therefore missing the availability of new data. These functions
need to read data from the internet, but are memoized so that the results are
reasonably up-to-date without having to make API calls repeatedly.
}
\examples{
\dontrun{
if(interactive()){
 cb_avail <- check_cb_avail()
 cb_avail |>
   dplyr::filter(program == "dec", vintage == 2020)
 }
}
\dontrun{
if(interactive()){
 qwi_avail <- check_qwi_avail()
 qwi_avail |>
   dplyr::filter(state_code == "09")
 }
}
}
\seealso{
\href{https://www.census.gov/data/developers/updates/new-discovery-tool.html}{US Census Bureau API Discovery Tool} \href{https://ledextract.ces.census.gov/}{LED Extraction Tool}
}
