% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ct5_clusters}
\alias{ct5_clusters}
\title{Five Connecticuts clusters}
\format{
A data frame with 169 rows and 2 variables:
\describe{
\item{town}{Town name}
\item{cluster}{Cluster label}
}
}
\source{
Levy, Don: Five Connecticuts 2010 Update. (2015). Produced for Siena College Research Institute and DataHaven based on original 1990 and 2000 designations from "Levy, Don, Orlando Rodriguez, and Wayne Villemez. 2004. The Changing Demographics of Connecticut - 1990 to 2000. Part 2: The Five Connecticuts. Storrs, Connecticut: University of Connecticut, The Connecticut State Data Center, Series, no. OP 2004-01."
}
\usage{
ct5_clusters
}
\description{
Cluster assignments for towns into "5 Connecticuts"--urban core, urban periphery, suburban, rural, and wealthy--based on median family income, poverty rate, and population density.
}
\seealso{
Other ref-datasets: 
\code{\link{acs_vars}},
\code{\link{decennial_vars}},
\code{\link{msa}},
\code{\link{naics_codes}},
\code{\link{neighborhood_tracts}},
\code{\link{occ_codes}},
\code{\link{proxy_pumas}},
\code{\link{regions}},
\code{\link{school_dists}},
\code{\link{table_nums}},
\code{\link{village2town}},
\code{\link{xwalk}},
\code{\link{zip2town}}
}
\concept{ref-datasets}
\keyword{datasets}
