% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwi-package.R
\docType{package}
\name{cwi-package}
\alias{cwi}
\alias{cwi-package}
\title{cwi: Prep for DataHaven's Community Wellbeing Index}
\description{
This is a collection of functions for collecting and analyzing Census data and other sources to write the 2019 Community Wellbeing Index reports
}
\author{
\strong{Maintainer}: Camille Seaberry \email{camille@ctdatahaven.org}

}
\keyword{internal}
