% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{decennial_vars}
\alias{decennial_vars}
\alias{decennial_vars10}
\title{Variable labels from the Decennial Census}
\format{
A data frame with 8959 rows (2010) or 9067 rows (2020) and 3 variables:
\describe{
\item{name}{Variable code containing the table number and variable number}
\item{label}{Readable label of each variable}
\item{concept}{Table name}
}
}
\source{
US Census Bureau via \code{tidycensus}
}
\usage{
decennial_vars

decennial_vars10
}
\description{
Dataset of Decennial Census variable labels, loaded from \code{tidycensus::load_variables()} and cleaned up slightly. 2010 and 2020 versions are saved separately and have different variable code formats. \code{decennial_vars} has an attribute \code{year} giving the year of the dataset.
}
\examples{
# get the year
attr(decennial_vars, "year")

}
\keyword{datasets}
\keyword{ref-datasets}
