% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_level_plot.R
\name{geo_level_plot}
\alias{geo_level_plot}
\title{Quickly make a \code{ggplot} to view data by geographic level}
\usage{
geo_level_plot(
  data,
  name = name,
  value = value,
  level = level,
  type = c("col", "hist", "point"),
  hilite = "dodgerblue",
  title = NULL,
  dark_gray = "gray20",
  light_gray = "gray60",
  ...
)
}
\arguments{
\item{data}{A data frame to use for plotting.}

\item{name}{Bare column name containing names, i.e. independent variable.}

\item{value}{Bare column name containing values, i.e. dependent variable.}

\item{level}{Bare column name containing geographic levels for fill.}

\item{type}{String: one of \code{"col"} (bar chart, using \code{\link[ggplot2:geom_bar]{ggplot2::geom_col()}}, \code{"point"} (dot plot, using \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}), or \code{"hist"} (histogram, using \code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}}); defaults \code{"col"}.}

\item{hilite}{String giving the highlight color, used for the lowest geography present.}

\item{title}{String giving the title, if desired, for the plot.}

\item{dark_gray}{String giving the named gray color for the highest geography; defaults \code{"gray20"}.}

\item{light_gray}{String giving the named gray color for the second lowest geography; defaults \code{"gray60"}.}

\item{...}{Any additional parameters to pass to the underlying geom function.}
}
\value{
A ggplot
}
\description{
This is a quick way to make a bar chart, a Cleveland dotplot, or a histogram from a set of data, filled by geographic level.
}
\seealso{
\code{\link[ggplot2:geom_bar]{ggplot2::geom_col()}} \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}} \code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}}
}
\keyword{quick-plotting-functions}
