% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gnh_tenure}
\alias{gnh_tenure}
\title{ACS demo data - tenure

This is a table of housing tenure data for Greater New Haven obtained with multi_geo_acs, to be used in testing and examples.}
\format{
A data frame with 45 rows and 5 variables:
\describe{
\item{\code{level}}{Geographic level}
\item{\code{name}}{Geography name}
\item{\code{tenure}}{Tenure: total households, owner-occupied, or renter-occupied}
\item{\code{estimate}}{Estimated count}
\item{\code{share}}{Calculated share of households, or \code{NA} for total}
}
}
\usage{
gnh_tenure
}
\description{
ACS demo data - tenure

This is a table of housing tenure data for Greater New Haven obtained with multi_geo_acs, to be used in testing and examples.
}
\keyword{datasets}
