% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenks.R
\name{jenks}
\alias{jenks}
\title{jenks}
\usage{
jenks(x, n = 5, true_jenks = FALSE, labels = NULL, ...)
}
\arguments{
\item{x}{A numeric vector to cut}

\item{n}{Number of bins, Default: 5}

\item{true_jenks}{Logical: should a "true" Jenks algorithm be used? If false, uses the faster Fisher-Jenks algorithm. See \code{classInt::classIntervals} docs for discussion. Default: FALSE}

\item{labels}{A string vector to be used as bin labels, Default: NULL}

\item{...}{Additional arguments passed on to \code{cut}}
}
\value{
A factor of the same length as x
}
\description{
Quickly cut a vector with the Jenks/Fisher algorithms
}
\details{
The Jenks/Fisher algorithms cut a set of numbers into bins in a way that highlights very high or very low values well. It's good to use for choropleths that need to convey imbalances or inequities.
}
\examples{
 set.seed(123)
 values <- rexp(30, 0.8)
 jenks(values, n = 4)
}
\seealso{
classInt::classIntervals
}
