% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenks.R
\name{jenks}
\alias{jenks}
\title{Quickly cut a vector with the Jenks/Fisher algorithms}
\usage{
jenks(x, n = 5, true_jenks = FALSE, labels = NULL, ...)
}
\arguments{
\item{x}{A numeric vector to cut}

\item{n}{Number of bins, Default: 5}

\item{true_jenks}{Logical: should a "true" Jenks algorithm be used? If false, uses the faster Fisher-Jenks algorithm. See \code{classInt::classIntervals} docs for discussion. Default: FALSE}

\item{labels}{A string vector to be used as bin labels, Default: NULL}

\item{...}{Additional arguments passed on to \code{\link[base:cut]{base::cut}}}
}
\value{
A factor of the same length as x
}
\description{
Given a numeric vector, this returns a factor of those values cut
into \code{n} number of breaks using the Jenks/Fisher algorithms. The algorithm(s) sets breaks in a way that highlights very high or very low values well. It's good to use for choropleths that need to convey imbalances or inequities.
}
\examples{
 set.seed(123)
 values <- rexp(30, 0.8)
 jenks(values, n = 4)
}
\seealso{
\code{\link[classInt:classIntervals]{classInt::classIntervals}}
}
