% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_decennial.R
\name{label_decennial}
\alias{label_decennial}
\title{Quickly add the labels of decennial variables}
\usage{
label_decennial(df, year = 2010, sumfile = "sf1", variable = variable)
}
\arguments{
\item{df}{A data frame/tibble.}

\item{year}{The year of decennial census data; defaults 2010.}

\item{sumfile}{A string: which summary file to use. Defaults to the 100 percent summary file (\code{"sf1"}), but can also be \code{"sf3"}.}

\item{variable}{The bare column name of variable codes; defaults to \code{variable}, as returned by \code{tidycensus::get_decennial}.}
}
\value{
A tibble
}
\description{
\code{tidycensus::get_decennial} returns a decennial data table with its variable codes, which can be joined with \code{cwi::decennial_vars10} to get readable labels. This function is just a quick wrapper around the common task of joining these two data frames.
}
\seealso{
\link{decennial_vars10}
}
