% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laus_trend.R
\name{laus_trend}
\alias{laus_trend}
\title{Fetch local area unemployment statistics (LAUS) data over time}
\usage{
laus_trend(
  names,
  startyear,
  endyear,
  measures = "all",
  annual = FALSE,
  key = Sys.getenv("BLS_KEY")
)
}
\arguments{
\item{names}{A character vector of place names to look up: as of now, this is limited to Connecticut, and its counties and towns.}

\item{startyear}{Numeric; first year of range}

\item{endyear}{Numeric; last year of range}

\item{measures}{A character vector of measures, containing any combination of \code{"unemployment rate"}, \code{"unemployment"}, \code{"employment"}, or \code{"labor force"}, or \code{"all"} (the default) as shorthand for all of the above.}

\item{annual}{Logical: whether to include annual averages along with monthly data. Defaults \code{FALSE}.}

\item{key}{A string giving the BLS API key. Defaults to the value in \code{Sys.getenv("BLS_KEY")}.}
}
\value{
A data frame / tibble with columns for the measure, area, year, month number and name, value, series ID, and footnotes returned from the API.
}
\description{
Fetch monthly LAUS data for a list of locations over a given time period, modeled after \code{blscrapeR::bls_api}. Requires a BLS API key.
}
\examples{
\dontrun{
laus_trend(c("Connecticut", "New Haven", "Hamden"), 2014, 2017, annual = TRUE)
}
}
