% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msa}
\alias{msa}
\title{Names and GEOIDs of regional MSAs}
\format{A data frame with 15 rows and 2 variables:
\describe{
\item{GEOID}{GEOID/FIPS code}
\item{name}{Name of MSA}
}}
\source{
US Census Bureau via \code{tidycensus}
}
\usage{
msa
}
\description{
A reference dataset of all the metropolitan statistical areas (MSAs) in New England.
}
\keyword{datasets}
