% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_geo_decennial.R
\name{multi_geo_decennial}
\alias{multi_geo_decennial}
\title{Fetch a decennial census table with multiple geography levels}
\usage{
multi_geo_decennial(
  table,
  year = 2010,
  neighborhoods = NULL,
  towns = "all",
  regions = NULL,
  counties = "all",
  state = "09",
  tracts = NULL,
  sumfile = "sf1",
  verbose = TRUE,
  key = NULL
)
}
\arguments{
\item{table}{A string giving the decennial census table number. These are generally formatted as one or more letters, 3 numbers, and optionally a letter.}

\item{year}{The year of the census table; currently defaults 2010 (most recent decennial census).}

\item{neighborhoods}{Temporarily deprecated: A named list of neighborhoods with their 11-digit tract GEOIDs (defaults NULL).}

\item{towns}{A character vector of towns to include; \code{"all"} (default) for all towns optionally filtered by county; or \code{NULL} to not fetch town-level table.}

\item{regions}{A named list of regions with their town names (defaults \code{NULL}).}

\item{counties}{A character vector of counties to include; \code{"all"} (default) for all counties in the state; or \code{NULL} to not fetch county-level table.}

\item{state}{A string: either name or two-digit FIPS code of a US state. Required; defaults \code{"09"} (Connecticut).}

\item{tracts}{A character vector of 11-digit FIPS codes of tracts to include, or \code{"all"} for all tracts optionally filtered by county. Defaults \code{NULL}.}

\item{sumfile}{A string giving the summary file to pull from. Defaults \code{"sf1"}; in some rare cases, \code{"sf3"} may be appropriate.}

\item{verbose}{Logical: whether to print summary of geographies included. Defaults \code{TRUE}.}

\item{key}{String: Census API key. If \code{NULL} (default), takes the value from \code{Sys.getenv("CENSUS_API_KEY")}.}
}
\value{
A tibble with GEOID, name, variable code, estimate, moe, geography level, state, and year, as applicable, for the chosen table.
}
\description{
Fetch a data table from the decennial census via \code{tidycensus} with your choice of geographies at multiple levels. For geographies made of aggregates, i.e. neighborhoods made of tracts or regions made of towns, the returned table will have estimates summed and margins of error calculated for the whole area.
}
\details{
This function essentially calls \code{tidycensus::get_decennial()} multiple times, depending on geographic levels chosen, and does minor cleaning, filtering, and aggregation. Note that the underlying \code{tidycensus::get_decennial()} requires a Census API key. As is the case with other \code{tidycensus} functions, \code{multi_geo_decennial} assumes this key is stored as \code{CENSUS_API_KEY} in your \code{.Renviron}. See \code{\link[tidycensus:census_api_key]{tidycensus::census_api_key()}} for installation.

Be advised that decennial table numbers may change from year to year, so if you're looking at trends, check FactFinder or another source to make sure the tables have the same meaning. Setting \code{verbose = TRUE} is helpful for this as well.
}
\examples{
\dontrun{
multi_geo_decennial("P001", 2010,
  neighborhoods = list(downtown = c("09009140100", "09009361401", "09009361402"),
    dixwell = "090091416"),
  towns = "all",
  regions = list(inner_ring = c("Hamden", "East Haven", "West Haven")),
  counties = "New Haven County")
}
}
\seealso{
\code{\link[tidycensus:census_api_key]{tidycensus::census_api_key()}}, \code{\link[tidycensus:get_decennial]{tidycensus::get_decennial()}}
}
