% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neighborhood_tracts}
\alias{neighborhood_tracts}
\alias{nhv_tracts}
\alias{nhv_bgrps}
\alias{bridgeport_tracts}
\alias{stamford_tracts}
\alias{hartford_tracts}
\title{City neighborhoods by tract or block group}
\format{A data frame
\describe{
\item{town}{For \code{hartford_tracts}, the name of the town, because both Hartford and West Hartford neighborhoods are included; otherwise, no \code{town} variable is needed}
\item{name}{Neighborhood name}
\item{geoid}{FIPS code of smallest geography. For all but \code{nhv_bgrps}, this is the 11-digit FIPS code of the Census tract; for \code{nhv_bgrps}, this is the 12-digit FIPS code of the block group.}
\item{tract}{6-digit FIPS code of the tract}
\item{block_group}{Single digit number of the block group, if applicable (\code{nhv_bgrps})}
\item{weight}{Share of tract/block group households in that neighborhood}
}}
\usage{
nhv_tracts

nhv_bgrps

bridgeport_tracts

stamford_tracts

hartford_tracts
}
\description{
Datasets of neighborhoods for New Haven (both Census tract and block group available), Hartford/West Hartford, Stamford, and Bridgeport. Some tracts cross between more than one neighborhood; use \code{weight} column for aggregating values such as populations.
}
\keyword{datasets}
