% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proxy_pumas}
\alias{proxy_pumas}
\title{Proxy PUMAs}
\format{
A list of 2 data frames, \code{county} and \code{cog}, with 19 and 54 rows, respectively, and 6 variables:
\describe{
\item{\code{puma}}{7-digit PUMA FIPS code}
\item{\code{region}}{Region name}
\item{\code{pop}}{Total population in the overlapping area between the region and the PUMA}
\item{\code{hh}}{Total households in the overlapping area between the region and the PUMA}
\item{\code{pop_weight}}{Population weight: share of the PUMA's population that's included in the region, to be used for population-based survey analysis}
\item{\code{hh_weight}}{Household weight: share of the PUMA's households that are included in the region, to be used for household-based survey analysis}
}
}
\source{
2021 & 2022 5-year ACS
}
\usage{
proxy_pumas
}
\description{
This is a list of 2 data frames giving PUMAs that make reasonable approximations of designated regions, with weights to apply to both population- and household-based measures. The data frame labeled \code{county} uses county-based PUMAs and 2021 ACS values; the data frame \code{cog} uses the new COG-based PUMAs and 2022 ACS values. When working with PUMS data or other weighted surveys, multiply the weights in the proxy table with the weights from the survey to account for how much of the PUMA overlaps the region.
}
\keyword{datasets}
