% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qwi_industry.R
\name{qwi_industry}
\alias{qwi_industry}
\title{Get employment counts and total payroll over time}
\usage{
qwi_industry(
  years,
  industries = cwi::naics_codes[["industry"]],
  state = "09",
  counties = NULL,
  annual = FALSE,
  key = NULL,
  retry = 5
)
}
\arguments{
\item{years}{A numeric vector of one or more years for which to get data}

\item{industries}{A character vector of NAICS industry codes; default is the 20 sectors plus "All industries" from the dataset \code{naics_codes}.}

\item{state}{A string of length 1 representing a state's FIPS code, name, or two-letter abbreviation; defaults to \code{"09"} for Connecticut}

\item{counties}{A character vector of county FIPS codes, or \code{"all"} for all counties. If \code{NULL} (the default), will return data just at the state level. For Connecticut, these now need to be COGs; data has been changed retroactively.}

\item{annual}{Logical, whether to return annual averages or quarterly data (default) .}

\item{key}{A Census API key. If \code{NULL}, defaults to the environmental variable \code{"CENSUS_API_KEY"}, as set by \code{tidycensus::census_api_key()}.}

\item{retry}{The number of times to retry the API call(s), since the server this comes from can be a bit finicky.}
}
\value{
A data frame / tibble
}
\description{
This gets data from the Quarterly Workforce Indicators (QWI) via the Census API. It's an alternative to \code{censusapi} that fetches a set of variables (employees and payroll) but makes a somewhat more dynamic API call. The API returns a maximum of 10 years of data; calling this function with more than 10 years will require multiple API calls, which takes a little longer.
}
\details{
Note that when looking at data quarterly, the payroll reported will be for that quarter, not the yearly payroll that you may be more accustomed to. As of November 2021, payroll data seems to be missing from the database; even the QWI Explorer app just turns up empty.
}
\examples{
\dontrun{
qwi_industry(2012:2017, industries = c("23", "62"), counties = "009")
}
}
\seealso{
Other fetching-functions: 
\code{\link{laus_trend}()},
\code{\link{multi_geo_acs}()},
\code{\link{multi_geo_decennial}()}
}
\concept{fetching-functions}
