% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qwi_industry.R
\name{qwi_industry}
\alias{qwi_industry}
\title{Get employees counts and total payroll over time}
\usage{
qwi_industry(years, industries = naics_codes$industry, counties,
  state = "09", annual = TRUE, key = Sys.getenv("CENSUS_API_KEY"))
}
\arguments{
\item{years}{A numeric vector of one or more years for which to get data}

\item{industries}{A character vector of NAICS industry codes; default is the 20 sectors plus "All industries" from the dataset \code{naics_codes}.}

\item{counties}{A character vector of county FIPS codes.}

\item{state}{A string of a state FIPS code; defaults to \code{"09"} for Connecticut}

\item{annual}{Logical, whether to return annual averages (default) or quarterly data.}

\item{key}{A Census API key. Defaults to the value at \code{"CENSUS_API_KEY"} in your \code{.Renviron} file, as set by \code{tidycensus::census_api_key}.}
}
\value{
A data frame / tibble
}
\description{
This gets data from the Quarterly Workforce Indicators (QWI) via the Census API. It's an alternative to \code{censusapi::getCensus} that fetches a set of variables (employees and payroll) but makes a somewhat more dynamic API call.
}
\examples{
\dontrun{
qwi_industry(2012:2017, industries = c("23", "62"), counties = "009")
}
}
