% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regions}
\alias{regions}
\title{Regions of Connecticut}
\format{
A named list of vectors, where names give the names of regions and vectors give the names of towns making up each region, including regional councils of governments.
}
\source{
DataHaven internal and CT OPM
}
\usage{
regions
}
\description{
A dataset of Connecticut regions by town
}
\keyword{datasets}
