% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_nonanswers.R
\name{sub_nonanswers}
\alias{sub_nonanswers}
\title{Remove non-answers and rescale percentage values}
\usage{
sub_nonanswers(
  .data,
  response = response,
  value = value,
  nons = c("Don't know", "Refused"),
  factor_response = TRUE
)
}
\arguments{
\item{.data}{A data frame}

\item{response}{Bare column name of where responses are found, including
those considered to be non-answers. Default: response}

\item{value}{Bare column name of values, Default: value}

\item{nons}{Character vector of responses to be removed.
Default: c("Don't know", "Refused")}

\item{factor_response}{Logical: if \code{TRUE} (default), returns response variable
as a factor. This is likely a more useful way to handle response
categories once non-answers have been removed.}
}
\value{
A data frame with the same number of columns as the original, but
fewer rows
}
\description{
This is a convenience function for removing what might be
considered non-answers ("don't know", "refused", etc.) and rescaling the
remaining values to add to 1.0.
}
\examples{
if (interactive()) {
xt <- system.file("extdata/test_xtab2018.xlsx", package = "cwi")
df <- read_xtabs(xt, process = TRUE) \%>\%
  dplyr::filter(code == "Q1") \%>\%
  sub_nonanswers()
}
}
