% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_nums}
\alias{table_nums}
\alias{basic_table_nums}
\alias{ext_table_nums}
\title{Common ACS table numbers}
\format{
A named list, where names correspond to abbreviated subjects and string values correspond to table numbers.
}
\usage{
basic_table_nums

ext_table_nums
}
\description{
Lists of ACS table numbers commonly used by DataHaven. \code{basic_table_nums} is used for making short profiles of towns, neighborhoods, and regions, while \code{ext_table_nums} is used for "extended" profiles, such as neighborhood profiles distributed as PDF on the DataHaven website.
}
\seealso{
Other ref-datasets: 
\code{\link{acs_vars}},
\code{\link{ct5_clusters}},
\code{\link{decennial_vars}},
\code{\link{msa}},
\code{\link{naics_codes}},
\code{\link{neighborhood_tracts}},
\code{\link{occ_codes}},
\code{\link{proxy_pumas}},
\code{\link{regions}},
\code{\link{school_dists}},
\code{\link{village2town}},
\code{\link{xwalk}},
\code{\link{zip2town}}
}
\concept{ref-datasets}
\keyword{datasets}
