% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/town_names.R
\name{town_names}
\alias{town_names}
\title{Clean up town names as returned from ACS}
\usage{
town_names(data, name_col)
}
\arguments{
\item{data}{A data frame}

\item{name_col}{Bare column name of town names}
}
\value{
A tibble/data frame with cleaned names and "not defined" towns removed
}
\description{
This function does two things: it removes text other than the town name from the column given as \code{name_col}, and it removes any rows for "county subdivisions not defined." For example, the string "Ansonia town, New Haven County, Connecticut" will become "Ansonia."
}
\examples{
pops <- tibble::tribble(
    ~name, ~total_pop,
    "County subdivisions not defined, New Haven County, Connecticut", 0,
    "Ansonia town, New Haven County, Connecticut", 18802,
    "Beacon Falls town, New Haven County, Connecticut", 6168,
    "Bethany town, New Haven County, Connecticut", 5513,
    "Branford town, New Haven County, Connecticut", 2802
)
town_names(pops, name_col = name)
}
\seealso{
Other utils: 
\code{\link{batch_csv_dump}()},
\code{\link{check_cb_avail}()},
\code{\link{fix_cogs}()},
\code{\link{jenks}()},
\code{\link{separate_acs}()}
}
\concept{utils}
