% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sf.R
\docType{data}
\name{tract_sf}
\alias{tract_sf}
\title{Connecticut tract shapefile}
\format{
Data frame of class \code{sf}.
\describe{
\item{name}{Tract FIPS code}
\item{geometry}{Tract geometry as \code{sfc_MULTIPOLYGON}}
}
}
\usage{
tract_sf
}
\description{
Data frame of Connecticut tracts with \code{sf} geometries.
}
\keyword{datasets}
