% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtab2df.R
\name{xtab2df}
\alias{xtab2df}
\title{Extract survey data and descriptions from crosstabs into a tidy data frame}
\usage{
xtab2df(.data, col = x1, code_pattern = "^[A-Z\\\\d_]{2,20}$")
}
\arguments{
\item{.data}{A data frame as returned from \code{read_xtab}.}

\item{col}{The bare column name of where to find question codes and text.
Default: x1, based on names assigned by \code{read_xtab}}

\item{code_pattern}{String: regex pattern denoting how to find cells that
contain only a question code, such as "Q10", "Q4B", or "ASTHMA". This is
pretty finicky, so you probably don't want to change it.
Default: \code{"^[A-Z\\\\d_]{2,20}$"}}
}
\value{
A data frame with the following columns:
\itemize{
\item code (if questions have codes in crosstabs)
\item q_number (if questions don't have codes in crosstabs, assigned in order they occur)
\item question
\item category (e.g. age, gender)
\item group (e.g. 18–34, male)
\item response
\item value
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Like \code{read_xtab} & \code{read_weights}, this is a bespoke function
to make it easier to extract data from the DataHaven Community Wellbeing
Survey. Applications to other crosstabs are probably limited unless their
formatting is largely the same. After reading a crosstab excel file, \code{xtab2df}
extracts the question codes (e.g. Q4A), question text, categories, and
demographic groups, and joins those descriptions with survey responses and
values, making it ready for analysis.
}
\examples{
if(interactive()){
  xtab <- read_xtabs(system.file("extdata/test_xtab2018.xlsx", package = "cwi"))
  xtab2df(xtab)
 }
}
\seealso{
\code{\link[=read_xtabs]{read_xtabs()}}
}
