% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xwalk}
\alias{xwalk}
\title{CT crosswalk}
\format{
A data frame with 67465 rows and 15 variables:
\describe{
\item{block}{Block FIPS code}
\item{block_grp}{Block group FIPS code}
\item{tract}{Tract FIPS code, based on county FIPS}
\item{tract_cog_fips}{Tract FIPS code, based on COG FIPS as of 2022 ACS}
\item{town}{Town name}
\item{town_fips}{Town FIPS code, based on county FIPS}
\item{town_cog_fips}{Town FIPS code, based on COG FIPS as of 2022 ACS}
\item{county}{County name}
\item{county_fips}{County FIPS code}
\item{cog}{COG}
\item{cog_fips}{COG FIPS code}
\item{msa}{Metro/micropolitan area name}
\item{msa_fips}{Metro/micropolitan area FIPS code}
\item{puma}{PUMA name}
\item{puma_fips}{PUMA FIPS code}
}
}
\source{
2020 and 2022 (for COGs & towns) TIGER shapefiles
}
\usage{
xwalk
}
\description{
A crosswalk between geographies in Connecticut, built off of TIGER shapefiles.
}
\keyword{datasets}
