% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xwalk}
\alias{xwalk}
\title{CT crosswalk}
\format{
A data frame with 67465 rows and 11 variables:
\describe{
\item{block}{Block FIPS code}
\item{block_grp}{Block group FIPS code}
\item{tract}{Tract FIPS code}
\item{town}{Town name}
\item{town_fips}{Town FIPS code}
\item{county}{County name}
\item{county_fips}{County FIPS code}
\item{msa}{Metro/micropolitan area name}
\item{msa_fips}{Metro/micropolitan area FIPS code}
\item{puma}{PUMA name}
\item{puma_fips}{PUMA FIPS code}
}
}
\source{
This is a filtered, cleaned, and pared down version of a file downloaded from \href{https://lehd.ces.census.gov/data/lodes/LODES7/ct/}{LEHD}, plus some spatial joins of TIGER shapefiles.
}
\usage{
xwalk
}
\description{
A shortened version of the CT crosswalk file from the LEHD/LODES files, to translate between geographies in Connecticut.
}
\keyword{datasets}
