% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zip2town}
\alias{zip2town}
\title{Zip to town lookup}
\format{
A data frame with 410 rows and 8 variables:
\describe{
\item{town}{Town name}
\item{zip}{5-digit zip code (ZCTA5)}
\item{inter_pop}{Population in this intersection of zips and towns}
\item{inter_hh}{Number of households in this intersection of zips and towns}
\item{pct_of_town_pop}{Percentage of the town's population that is also in this zip}
\item{pct_of_town_hh}{Percentage of the town's households that are also in this zip}
\item{pct_of_zip_pop}{Percentage of the zip's population that is also in this town}
\item{pct_of_zip_hh}{Percentage of the zip's households that are also in this town}
}
}
\source{
Cleaned-up version of the Census \href{https://www2.census.gov/geo/docs/maps-data/data/rel2022/acs22_cousub22_zcta520_st09.txt}{2022 ZCTA to county subdivision relationship file}, updated for Connecticut's 2022 revisions from counties to COGs.
}
\usage{
zip2town
}
\description{
A crosswalk of Connecticut's ZCTA5s and towns with shares of populations and households in overlapping areas. Each row corresponds to a combination of zip and town; therefore, some zips have more than one observation, as do towns.
}
\seealso{
Other ref-datasets: 
\code{\link{acs_vars}},
\code{\link{ct5_clusters}},
\code{\link{decennial_vars}},
\code{\link{msa}},
\code{\link{naics_codes}},
\code{\link{neighborhood_tracts}},
\code{\link{occ_codes}},
\code{\link{proxy_pumas}},
\code{\link{regions}},
\code{\link{school_dists}},
\code{\link{table_nums}},
\code{\link{village2town}},
\code{\link{xwalk}}
}
\concept{ref-datasets}
\keyword{datasets}
